[SOIN] Extrais un composant "Modale"
... et l'utilise pour la modale de détails d'une mesure.
